run server tests on python 312
