+++ .github/dependabot.yml
@@ -9,7 +9,18 @@ updates:
     directory: "/" # Location of package manifests
     schedule:
       interval: "weekly"
+
+  - package-ecosystem: "npm" # See documentation for possible values
+    directory: "/e2e/" # Location of package manifests
+    schedule:
+      interval: "weekly"
+
   - package-ecosystem: "pip" # See documentation for possible values
     directory: "/" # Location of package manifests
+    schedule:
+      interval: "weekly"
+
+  - package-ecosystem: "github-actions"
+    directory: "/"
     schedule:
       interval: "weekly"
