+++ .github/workflows/ci-server.yml
@@ -7,7 +7,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ['3.10']
+        python-version: ['3.10', '3.12']
     env:
       INSTALL_PY_MODULES: true
       RUN_SERVICES: true
@@ -29,7 +29,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ['3.10']
+        python-version: ['3.10', '3.12']
     env:
       INSTALL_PY_MODULES: true
       RUN_SERVICES: true
@@ -55,7 +55,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ['3.10']
+        python-version: ['3.10', '3.12']
     env:
       INSTALL_PY_MODULES: true
       INSTALL_PY_EDITABLE: true
