+++ e2e/server/core-requirements.in
@@ -1,2 +1,2 @@
-honcho==1.0.1
+honcho
 git+https://github.com/superdesk/superdesk-core.git@release/3.1#egg=superdesk-core
