+++ server/requirements.txt
@@ -23,4 +23,4 @@ aioresponses
 
 -e .
 # Install in editable state so we get feature fixtures
--e git+https://github.com/superdesk/superdesk-core.git@develop#egg=superdesk-core
+-e git+https://github.com/superdesk/superdesk-core.git@release/3.1#egg=superdesk-core
